Create first tests and the hit, isSunk methods
The factory function for the ships contains at the moment the current
length, which represents the health of the ship, as well as the methods
to identify when the ship has been hit, which decrements the ship's
health. On the other hand, the isSunk method evaluates the value of the
currentHealth, if it hits 0 then the ship is sunk.

The testScript file was created to replace the index.js file for the
moment, as the import style.css wreaks havoc by trying to run the tests
from the input file.

In the testScript.test.js two describe functions were used to limit
the scope for the ship 1 and ship 2, in which it has been evaluated
so far that the hit method returns the coordinate that has been hit, and
the isSunk method, that returns the state of the ship.