Create gameboard factory function
At the moment only the first board is created, which contains a method
capable of creating random positions for the 10 ships.

This method passes in the desired length of the ships, where the first
3 ships have a length of 3, the next 3 a length of 2, and the last 4 a
length of 1.

At the outset a ship is created without coordinates and pushed to the
ships array of the gameboard. Subsequently, with the help of do while
the random coordinates are created according to the desired ship's
length. Then, if that location does not contain an already occupied
coordinate, then it is returned and assigned to the last ship in the
ships array.