Fix bugs and style gamboards
The styleGridPlayer1 and 2 were added to the factory function for the
gameboard, which selects the respective gameboards according to the
ship occupying the position to add the ship class, which sets the
background color to a blue hue.

On the other hand, in the HTML file the class for selecting the
gameboard grids begins at 0, as this values was not present but the
generator of coordinates takes into consideration such value, leading
to random errors when the ships were created.