Create playerFactory and receiveAttack method in gameboardFactory
The playerFactory contains at the moment a single method, which is
attack() that takes as a string the player that is being attacked.

This function leverages an eventlistener to select any grid from the
opponent's gameboard and passes the information to the receiveAttack
method from the gameboardFactory. The DOM manipulation will be put in
a different module later.

The receiveAttack method evaluates that the grid selected does not
contain the data attribute set to distinguish which ship was selected
and that the current grid does not have the empty class on it, otherwise
the current grid will be added that class to change the background
color to red to indicate that the player failed to attack a ship.