Creat gameFlow object
In this object all of the objects for the game are created and their
methods are used to allow the players to make their moves in their
respective turns. The gameFlow object has the function gameLoop, which
enables to carry out the game and change the turn according to the move
played by the current player.

On the other hand, the object view was created, which oversees all the
changes in the display, such as displaying the misses or hits based on
the player's moves.