Create game loop
The loop was created with the help of a recursive function and the
requestAnimationFrame function. At the outset, the event listener for
player1 was declared within this loop, which triggered the event
listener multiple times, which made it impossible to evaluate whether
the player had reselected a grid.

This was solved by placing the event listener outside the loop and
executing the callback function only if the current turn carries the
value of "player1".

On the other hand, the function checkIfRepeatTurn defined in the
gambeboard factory was renamed to "setNextTurn" and moved to the
gameFlow object in order to mutate the currenTurn variable, which
dictates whose turn is to play.