Allow players to move again if they hit a ship
Once the grid was selected and given the corresponding class to
highlight the result of the player's move, the added class of that grid
is evaluated to see if it contains the class "hit", which indicates that
a ship was hit. When that condition is met the current player can move
again, but also the ship's hit method is called to decreased the health
of the respective ship. Likewise, the isSunk method is called afterwards
to evaluate whether the ship is sunk or not.

When a ship is sunk it is given the class sunk, which changes the
background color to black.