Add logic to end game
Once all the ships in one of the borders are sunk, the game loop is
over. The logic remains to be separated from the current module.

Likewise, the event listener for the second board was modified to
add the event listener to each grid and not to the whole board, as it
led to the bug of marking the whole board if the player clicked several
times on it.