Simplify code and remove global variables
Parts of the code were stored in functions to make it more readable.
Furthermore, the global variable "winner" used to command whether the
game should end or not is now defined within the gameFlow object. At the
outset this was not done as the property "winner" of the object was not
updated with the corresponding value outside the object, but only within
it.

The above problem was solved by defining a getter, which returns the
current value of the property.